helloapp/requirements/test.txt
-r runtime.txt
attrs==19.3.0
entrypoints==0.3
importlib-metadata==1.5.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-flask==0.15.1
wcwidth==0.1.8
zipp==2.1.0
